Updated migrations and added if exists check

+++ database/migrations/2025_04_23_111700_create_page_data_table.php
@@ -11,20 +11,22 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('page_data', function (Blueprint $table) {
-            $table->unsignedInteger('data_id', true);
-            $table->primary('data_id');
-            $table->unsignedInteger('page_id');
-            $table->string('page_key')->nullable();
-            $table->string('page_module')->nullable();
-            $table->integer('page_sequence')->nullable();
-            $table->text('page_data')->nullable();
-            $table->dateTime('create_date')->nullable();
-            $table->timestamp('modified_date')->useCurrentOnUpdate()->nullable()->useCurrent();
+        if (! Schema::hasTable('page_data')) {
+            Schema::create('page_data', function (Blueprint $table) {
+                $table->unsignedInteger('data_id', true);
+                $table->primary('data_id');
+                $table->unsignedInteger('page_id');
+                $table->string('page_key')->nullable();
+                $table->string('page_module')->nullable();
+                $table->integer('page_sequence')->nullable();
+                $table->text('page_data')->nullable();
+                $table->dateTime('create_date')->nullable();
+                $table->timestamp('modified_date')->useCurrentOnUpdate()->nullable()->useCurrent();
 
-            // foreign keys
-            $table->foreign('page_id')->references('page_id')->on('pages')->onDelete('RESTRICT')->onUpdate('RESTRICT');
-        });
+                // foreign keys
+                $table->foreign('page_id')->references('page_id')->on('pages')->onDelete('RESTRICT')->onUpdate('RESTRICT');
+            });
+        }
     }
 
     /**
