+++ database/migrations/2025_04_23_111700_create_page_widgets_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('page_widgets', function (Blueprint $table) {
-            $table->unsignedInteger('widget_id', true);
-            $table->primary('widget_id');
-            $table->unsignedInteger('page_id');
-            $table->string('page_key')->nullable();
-            $table->string('widget_key')->nullable();
-            $table->string('widget_value')->nullable();
-            $table->enum('widget_active', ['0', '1', '2'])->nullable()->default('0');
-            $table->integer('widget_order')->nullable();
-            $table->dateTime('create_date')->nullable();
-            $table->timestamp('modified_date')->useCurrentOnUpdate()->nullable()->useCurrent();
+        if (! Schema::hasTable('page_widgets')) {
+            Schema::create('page_widgets', function (Blueprint $table) {
+                $table->unsignedInteger('widget_id', true);
+                $table->primary('widget_id');
+                $table->unsignedInteger('page_id');
+                $table->string('page_key')->nullable();
+                $table->string('widget_key')->nullable();
+                $table->string('widget_value')->nullable();
+                $table->enum('widget_active', ['0', '1', '2'])->nullable()->default('0');
+                $table->integer('widget_order')->nullable();
+                $table->dateTime('create_date')->nullable();
+                $table->timestamp('modified_date')->useCurrentOnUpdate()->nullable()->useCurrent();
 
-            // foreign keys
-            $table->foreign('page_id')->references('page_id')->on('pages')->onDelete('RESTRICT')->onUpdate('RESTRICT');
-        });
+                // foreign keys
+                $table->foreign('page_id')->references('page_id')->on('pages')->onDelete('RESTRICT')->onUpdate('RESTRICT');
+            });
+        }
     }
 
     /**
